Made some changes to what supervisors see
Allowed supervisors to see the logbooks from which students want a log
entry approved, the url for the logbooks depend on a migration currently
in the master branch which includes a slug field in the logbook model.
Could even load the log entries directly into the page and have drop
downs which have the entry information requiring approval, will discuss.